refactor: stop using a SharedWorker
Viewers are not refreshed periodically anymore, only when a connection is created or edited. Sending telemetry has been moved to a WebWorker.
