+++ apps/webapp/src/vite-env.d.ts
@@ -1,5 +1,4 @@
 /// <reference types="vite/client" />
-/// <reference types="vite-plugin-comlink/client" />
 
 interface ImportMetaEnv {
   readonly VITE_COMMIT_SHA?: string;
